crypto/totient/src/requirements.txt
pycryptodome
gmpy2